refactored db.py, eliminated unused modules, and added documentation
